+++ lib/qiita/markdown/summary_processor.rb
@@ -16,11 +16,12 @@ module Qiita
       def self.default_filters
         [
           Filters::Greenmat,
+          Filters::UserInputSanitizer,
           Filters::Simplify,
           Filters::Emoji,
           Filters::Mention,
           Filters::ExternalLink,
-          Filters::Sanitize,
+          Filters::FinalSanitizer,
           Filters::Truncate,
         ]
       end