+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## 0.19.0
+
 - Drop 2.0 and 2.1 from support Ruby versions
 - Rename `Sanitize` as `FinalSanitizer`
 - Add `:strict` context for stricter sanitization